moved one_hot and metrics to utils
