+++ docs/source/metrics.rst
@@ -1,7 +1,7 @@
-torchgeometry.metrics
-=====================
+torchgeometry.utils.metrics
+===========================
 
-.. currentmodule:: torchgeometry.metrics
+.. currentmodule:: torchgeometry.utils.metrics
 
 .. autofunction:: confusion_matrix
 .. autofunction:: mean_iou
