+++ docs/source/utils.rst
@@ -6,3 +6,9 @@ torchgeometry.utils
 .. autofunction:: tensor_to_image
 .. autofunction:: image_to_tensor
 .. autofunction:: create_meshgrid
+.. autofunction:: one_hot
+
+.. toctree::
+   :maxdepth: 2
+
+   metrics
