+++ mypy_files.txt
@@ -7,6 +7,11 @@ torchgeometry/contrib/extract_patches.py
 torchgeometry/contrib/max_blur_pool.py
 torchgeometry/feature/harris.py
 torchgeometry/feature/nms.py
+torchgeometry/filters/blur.py
+torchgeometry/filters/gaussian.py
+torchgeometry/filters/laplacian.py
+torchgeometry/filters/median.py
+torchgeometry/filters/sobel.py
 torchgeometry/geometry/homography_warper.py
 torchgeometry/geometry/depth_warper.py
 torchgeometry/geometry/pinhole.py
@@ -14,15 +19,13 @@ torchgeometry/geometry/imgwarp.py
 torchgeometry/geometry/crop.py
 torchgeometry/geometry/affine_warp.py
 torchgeometry/geometry/pyramid.py
-torchgeometry/filters/gaussian.py
-torchgeometry/filters/laplacian.py
-torchgeometry/filters/sobel.py
-torchgeometry/filters/blur.py
-torchgeometry/filters/median.py
 torchgeometry/losses/ssim.py
 torchgeometry/losses/dice.py
 torchgeometry/losses/focal.py
 torchgeometry/losses/tversky.py
 torchgeometry/losses/depth_smooth.py
-torchgeometry/metrics/confusion_matrix.py
-torchgeometry/metrics/mean_iou.py
+torchgeometry/utils/grid.py
+torchgeometry/utils/image.py
+torchgeometry/utils/one_hot.py
+torchgeometry/utils/metrics/confusion_matrix.py
+torchgeometry/utils/metrics/mean_iou.py
