+++ torchgeometry/losses/__init__.py
@@ -1,7 +1,7 @@
-from .one_hot import one_hot
 from .ssim import SSIM, ssim
 from .dice import DiceLoss, dice_loss
 from .tversky import TverskyLoss, tversky_loss
 from .focal import FocalLoss, focal_loss
-from .depth_smooth import InverseDepthSmoothnessLoss
-from .depth_smooth import inverse_depth_smoothness_loss
+from .depth_smooth import (
+    InverseDepthSmoothnessLoss, inverse_depth_smoothness_loss
+)
