+++ torchgeometry/utils/metrics/__init__.py
@@ -1,2 +1,7 @@
 from .confusion_matrix import confusion_matrix
 from .mean_iou import mean_iou
+
+__all__ = [
+    "confusion_matrix",
+    "mean_iou",
+]
